Improve several things
This commit contains several changes which updates the "Places" page as
well as the GPS Navigator and some other parts.
At first, common functionalities of both pages has been merged in the
JavaScript classes "LocalCoordinateStore" and "Uplink".
Another important modification is the conversion of
"GPSNavigationController" and "GPS NavigationController", to regular
classes. This was necessary to illustrate the dependencies between all
these classes.
Furthermore this commit fixes multiple bugs and improves many features,
for more details take a look at the changes.